Fixed nodeJS compatibility
